+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Change log
 
 ### Release 1.3.5
+**KPCLib**
+ - Added `Exists()` in `ProtectedBinaryDictionar`
+
 **PassXYZLib**
- - Added SearchEntries()
+ - Added `SearchEntries()`
 
 ### Release 1.3.4
  - Bug fixes