+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Release 1.3.5
 **KPCLib**
- - Added `Exists()` in `ProtectedBinaryDictionar`
+ - Added `Exists()` in `ProtectedBinaryDictionary`
 
 **PassXYZLib**
  - Added `SearchEntries()`